Remove some bloat (particularly from process_attributes), only show price in certain contexts
process_attributes was starting to gain bloat. I may need to start
cleaning this up again in the near future.

generate_item_url was a little shoddy in its reliance on custom
attributes, so an override parameter for mode was added instead.

Added geneterate_item_price_string for generating a proper line
for use in attributes and such.