Make tag and gift icons 16x16, make item tool bar position proportional to backpack pages, better player badge positioning, fix rare dialog open error
Item cells are now the same innter size as images, for best compatibility
and flow.

Player name and status are now under the portrait

ID bug was related to JS's handling of 64 bit integers and occasional
games that were a very high integer causing it to open the wrong ones
or not at all. And this is why casting to int should only be used
in known math operations.