Make schema/asset server downtime less of a point of failure, add CacheError hierarchy
Adding a root CacheError exception with the first child class being
CacheEmptyError for when no usable object can be found to return
for cached calls.

Also split up asset, schema, and schema-using calls so that there is
a better chance of them each working when the other goes down.

Also cleaned up schema display filtering dance so that
the proper lists get set after or before filtering as needed.

/Should/ finally fix the issue of ACO function trying to return
non-filtered objects on errors