Simplify doctype, remove unused max/min attr values
All browsers all the way back to IE6 go into standards mode when
they see the current HTML5 doctype. Which is the main reason for
having one, most doctypes are not very different from each other
or are exactly the same. Therefore using the HTML5 doctype is
clearly the best route.

Valve is starting to remove the minimum/maximum attr value fields.
So the description string substitutions now have a <type> instead.